+++ pkg/services/alerting/notifier_test.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 
 	"github.com/grafana/grafana/pkg/models"
+	m "github.com/grafana/grafana/pkg/models"
 	. "github.com/smartystreets/goconvey/convey"
 )
 
@@ -23,7 +24,7 @@ func (fn *FakeNotifier) NeedsImage() bool {
 
 func (fn *FakeNotifier) Notify(alertResult *EvalContext) {}
 
-func (fn *FakeNotifier) MatchSeverity(result models.AlertSeverityType) bool {
+func (fn *FakeNotifier) PassesFilter(rule *Rule) bool {
 	return fn.FakeMatchResult
 }
 
@@ -34,7 +35,7 @@ func TestAlertNotificationExtraction(t *testing.T) {
 			ctx := &EvalContext{
 				Firing: false,
 				Rule: &Rule{
-					Severity: models.AlertSeverityCritical,
+					State: m.AlertStateAlerting,
 				},
 			}
 			notifier := &FakeNotifier{FakeMatchResult: false}
@@ -42,12 +43,12 @@ func TestAlertNotificationExtraction(t *testing.T) {
 			So(shouldUseNotification(notifier, ctx), ShouldBeTrue)
 		})
 
-		Convey("exeuction error cannot be ignored", func() {
+		Convey("execution error cannot be ignored", func() {
 			ctx := &EvalContext{
 				Firing: true,
 				Error:  fmt.Errorf("I used to be a programmer just like you"),
 				Rule: &Rule{
-					Severity: models.AlertSeverityCritical,
+					State: m.AlertStateOK,
 				},
 			}
 			notifier := &FakeNotifier{FakeMatchResult: false}
@@ -59,7 +60,7 @@ func TestAlertNotificationExtraction(t *testing.T) {
 			ctx := &EvalContext{
 				Firing: true,
 				Rule: &Rule{
-					Severity: models.AlertSeverityCritical,
+					State: models.AlertStateAlerting,
 				},
 			}
 			notifier := &FakeNotifier{FakeMatchResult: true}
@@ -70,9 +71,7 @@ func TestAlertNotificationExtraction(t *testing.T) {
 		Convey("firing alert that dont match", func() {
 			ctx := &EvalContext{
 				Firing: true,
-				Rule: &Rule{
-					Severity: models.AlertSeverityCritical,
-				},
+				Rule:   &Rule{State: m.AlertStateOK},
 			}
 			notifier := &FakeNotifier{FakeMatchResult: false}