+++ public/app/features/alerting/alert_tab_ctrl.ts
@@ -17,7 +17,6 @@ export class AlertTabCtrl {
   alert: any;
   conditionModels: any;
   evalFunctions: any;
-  severityLevels: any;
   noDataModes: any;
   addNotificationSegment;
   notifications;
@@ -41,7 +40,6 @@ export class AlertTabCtrl {
     this.subTabIndex = 0;
     this.evalFunctions = alertDef.evalFunctions;
     this.conditionTypes = alertDef.conditionTypes;
-    this.severityLevels = alertDef.severityLevels;
     this.noDataModes = alertDef.noDataModes;
     this.appSubUrl = config.appSubUrl;
   }
@@ -70,7 +68,7 @@ export class AlertTabCtrl {
       this.notifications = res;
 
       _.each(this.alert.notifications, item => {
-        var model = _.findWhere(this.notifications, {id: item.id});
+        var model = _.find(this.notifications, {id: item.id});
         if (model) {
           model.iconClass = this.getNotificationIcon(model.type);
           this.alertNotifications.push(model);
@@ -88,7 +86,7 @@ export class AlertTabCtrl {
   }
 
   getAlertHistory() {
-    this.backendSrv.get(`/api/alert-history?dashboardId=${this.panelCtrl.dashboard.id}&panelId=${this.panel.id}`).then(res => {
+    this.backendSrv.get(`/api/annotations?dashboardId=${this.panelCtrl.dashboard.id}&panelId=${this.panel.id}&limit=50`).then(res => {
       this.alertHistory = _.map(res, ah => {
         ah.time = moment(ah.timestamp).format('MMM D, YYYY HH:mm:ss');
         ah.stateModel = alertDef.getStateDisplayModel(ah.newState);
@@ -125,12 +123,16 @@ export class AlertTabCtrl {
   }
 
   notificationAdded() {
-    var model = _.findWhere(this.notifications, {name: this.addNotificationSegment.value});
+    var model = _.find(this.notifications, {name: this.addNotificationSegment.value});
     if (!model) {
       return;
     }
 
-    this.alertNotifications.push({name: model.name, iconClass: this.getNotificationIcon(model.type)});
+    this.alertNotifications.push({
+      name: model.name,
+      iconClass: this.getNotificationIcon(model.type),
+      isDefault: false
+    });
     this.alert.notifications.push({id: model.id});
 
     // reset plus button
@@ -155,8 +157,7 @@ export class AlertTabCtrl {
       alert.conditions.push(this.buildDefaultCondition());
     }
 
-    alert.noDataState = alert.noDataState || 'unknown';
-    alert.severity = alert.severity || 'critical';
+    alert.noDataState = alert.noDataState || 'no_data';
     alert.frequency = alert.frequency || '60s';
     alert.handler = alert.handler || 1;
     alert.notifications = alert.notifications || [];
@@ -317,11 +318,6 @@ export class AlertTabCtrl {
   }
 
   evaluatorParamsChanged() {
-    ThresholdMapper.alertToGraphThresholds(this.panel);
-    this.panelCtrl.render();
-  }
-
-  severityChanged() {
     ThresholdMapper.alertToGraphThresholds(this.panel);
     this.panelCtrl.render();
   }