+++ public/app/plugins/datasource/grafana/datasource.ts
@@ -12,6 +12,16 @@ class GrafanaDatasource {
       maxDataPoints: options.maxDataPoints
     });
   }
+
+  annotationQuery(options) {
+    return this.backendSrv.get('/api/annotations', {
+      from: options.range.from.valueOf(),
+      to: options.range.to.valueOf(),
+      limit: options.limit,
+      type: options.type,
+    });
+  }
+
 }
 
 export {GrafanaDatasource};